# Fix performance downgrade with Mail::Utilities.to_crlf/to_lf