Fixes Issue: https://github.com/mikel/mail/issues/1118.
Reverts: https://github.com/mikel/mail/pull/956 and https://github.com/mikel/mail/pull/961.

Test suite time on my local machine before and after this change:

#### Ruby 2.4.1

Before:

```bash
$ be rspec spec/
Running Specs under Ruby Version 2.4.1
Running Specs for Mail Version 2.8.0.edge

Finished in 2.43 seconds (files took 0.40158 seconds to load)
1762 examples, 0 failures, 5 pending
```

After:

```bash
$ be rspec spec/
Running Specs under Ruby Version 2.4.1
Running Specs for Mail Version 2.8.0.edge

Finished in 1.55 seconds (files took 0.41858 seconds to load)
1760 examples, 0 failures, 5 pending
```

#### Ruby 2.1.1

Before:

```bash
$ be rspec spec/
Running Specs under Ruby Version 2.1.1
Running Specs for Mail Version 2.8.0.edge

Finished in 2.65 seconds (files took 0.3783 seconds to load)
1762 examples, 0 failures, 5 pending
```

After:

```bash
$ be rspec spec/
Running Specs under Ruby Version 2.1.1
Running Specs for Mail Version 2.8.0.edge

Finished in 1.75 seconds (files took 0.4047 seconds to load)
1760 examples, 0 failures, 5 pending
```

Or, you can compare Travis specs run times [Ruby 2.4 - before](https://travis-ci.org/mikel/mail/jobs/296108104) (3.59 seconds) vs [Ruby 2.4 - after](https://travis-ci.org/mikel/mail/jobs/296130765) (2.52 seconds)