@jeremy 2017-11-02
Backported to 2-7-stable for 2.7.1 release.